+++ app/Http/Controllers/AuthController.php
@@ -44,31 +44,4 @@ class AuthController extends Controller
             ]);
         }
     }
-
-    public function confirmUser(request $request, $token) {
-        $currentUrl = $request->fullUrl();
-        $user = Auth::user();
-        $user_confirmation = UserConfirmation::where('confirmation_link', $currentUrl)->first();
-
-        if ($user->confirmed) {
-            return redirect('/')->with('info', 'Your account is already confirmed.');
-        }
-
-        if (!$user_confirmation || $user->email !== $user_confirmation->email) {
-            return redirect('/')->with('error', 'The link is invalid.');
-        }
-
-        if (Carbon::now()->greaterThan($user_confirmation->expired_at)) {
-            return view('auth.expired-page')->with('error', 'The link has expired.');
-        }
-
-        $current_user = User::find($user->id);
-        $current_user->update([
-            'confirmed' => true
-        ]);
-
-        $user_confirmation->delete();
-        
-        return redirect('/')->with('success', 'Your email has been confirmed!');
-    }
 }