+++ resources/views/user/sended-confirmation-user.blade.php
@@ -1,7 +1,7 @@
 @extends('layouts.app')
 @section('title', 'Market - Confirmation User')
 @section('content')
-@vite('resources/css/auth/sended-confirmation-user.css')
+@vite('resources/css/user/sended-confirmation-user.css')
 
 <p id="sended-confirmation-user-title">We've sent a confirmation email to your inbox.</p>
 <p id="sended-confirmation-user-description">Click the link inside to verify your account.</p>